+++ distage/distage-testkit-core/src/main/scala/izumi/distage/testkit/services/dstest/TestRegistration.scala
@@ -1,7 +1,7 @@
 package izumi.distage.testkit.services.dstest
 
 import izumi.distage.model.providers.Functoid
-import izumi.distage.testkit.services.dstest.DistageTestRunner.TestId
+import izumi.distage.testkit.services.dstest.model.TestId
 import izumi.fundamentals.platform.language.SourceFilePosition
 
 trait TestRegistration[F[_]] {
