+++ distage/distage-testkit-scalatest/src/main/scala/org/scalatest/distage/DistageScalatestReporter.scala
@@ -1,7 +1,8 @@
 package org.scalatest.distage
 
 import izumi.distage.model.exceptions.runtime.IntegrationCheckException
-import izumi.distage.testkit.services.dstest.DistageTestRunner.{SuiteData, TestMeta, TestReporter, TestStatus}
+import izumi.distage.testkit.services.dstest.DistageTestRunner.{TestReporter, TestStatus}
+import izumi.distage.testkit.services.dstest.model.{SuiteData, TestMeta}
 import izumi.distage.testkit.services.scalatest.dstest.DistageTestsRegistrySingleton
 import izumi.fundamentals.platform.strings.IzString._
 import org.scalatest.Suite.{getIndentedTextForInfo, getIndentedTextForTest}
